% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CombineTimeTables}
\alias{CombineTimeTables}
\title{combine time tables}
\usage{
CombineTimeTables(single_timetable, group_timetable, join_type = c("cbind"))
}
\arguments{
\item{single_timetable}{a new timetable returned from 'PutOnTimeTable'}

\item{group_timetable}{a group of previously combined time tables or a new time table returned from 'PutOnTimeTable'}

\item{join_type}{"right join" or "cbind". "cbind" is more efficient but assumes everything is ordered properly.}
}
\description{
combine time tables
}
