% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetTimeTable}
\alias{GetTimeTable}
\title{creates a uniformly-spaced discrete-time grid data.table}
\usage{
GetTimeTable(
  start_times = NULL,
  end_times = NULL,
  IDs = NULL,
  sample_interval_minute = 30
)
}
\arguments{
\item{start_times}{vector of date times denoting the 'start' of a time series (e.g. the start of a patient's stay in the ICU)}

\item{end_times}{vector of date times denoting the 'end' of a time series  (e.g. the end of a patient's stay in the ICU)}

\item{IDs}{vector of ID associated with each discrete time series. If null, each start/end time pair is given a unique ID.}

\item{sample_interval_minute:}{numeric scalar, the time elapsed between one grid point to another (in minutes).}
}
\value{
long tables of variables, time points at every \code{sample_interval_minute} from  start_times to end_times
}
\description{
creates a uniformly-spaced discrete-time grid data.table
}
