% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetTimeTableTemplate}
\alias{GetTimeTableTemplate}
\title{creates a uniformly-spaced discrete-time grid data.table on which all other variables are appended to.}
\usage{
GetTimeTableTemplate(
  start_times = NULL,
  end_times = NULL,
  IDs = NULL,
  sample_interval = "30 minutes"
)
}
\arguments{
\item{start_times}{vector of date times denoting the 'start' of a time series (e.g. the start of a patient's stay in the ICU)}

\item{end_times}{vector of date times denoting the 'end' of a time series  (e.g. the end of a patient's stay in the ICU)}

\item{IDs}{vector of ID associated with each discrete time series. If null, each start/end time pair is given a unique ID.}

\item{sample_interval:}{a string that denotes the time (and the respective units) between each index in the TimeTable. All meaningful specifications in the English language are supported. Some example of valid inputs: '45 secs', '30 min', '22.2 mins', '2 minutes', or '3 years'.}
}
\value{
long table of variables, time points at every \code{sample_interval} from  start_times to end_times
}
\description{
creates a uniformly-spaced discrete-time grid data.table on which all other variables are appended to.
}
