% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Merge2TimeTables}
\alias{Merge2TimeTables}
\title{merge time tables}
\usage{
Merge2TimeTables(giver_tt, reciever_tt, join_type = c("right_join"))
}
\arguments{
\item{giver_tt}{a new TimeTable returned from 'PutOnTimeTable'}

\item{reciever_tt}{a group of previously combined time tables or a new time table returned from 'PutOnTimeTable'}

\item{join_type}{"right join", "cbind", "cbind_fast". "cbind_fast" is slightly more efficient but assumes everything is ordered properly, while "cbind" orders
your data.table by IDs and grid_datetimes. "right_join" is the safest but least efficient and will even work with TimeTables of unequal rows.}
}
\description{
merge time tables
}
