% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PutOnTimeTable}
\alias{PutOnTimeTable}
\title{Put Variable on time table}
\usage{
PutOnTimeTable(
  values,
  IDs,
  datetimes = NULL,
  var_name = NULL,
  duration_hr = NA,
  sample_interval_minute = 30,
  aggregration_type = "latest",
  grid = NULL,
  start_times = NULL,
  end_times = NULL,
  IDs_times = NULL
)
}
\arguments{
\item{values}{vector of values for grid}

\item{IDs}{vector of associated IDs for each value and datatimes}

\item{datetimes}{vector of date times associate with each values / IDs}

\item{var_name}{scalar name of variable}

\item{duration_hr}{how many hours a variable's valuable can be carried forward if proceeding measurements are missing.}

\item{sample_interval_minute}{scalar time between one time grid point and another (in minutes)}

\item{aggregration_type}{how to handle multiple measurements for same time point; 'latest' or 'mean'}

\item{grid}{grid precomputed from GetTimeTable()}

\item{start_times}{optional argument used if 'grid' is NULL vector of date times denoting the 'end' of a time series.}

\item{end_times}{optional argument used if 'grid' is NULL, vector of date times denoting the 'start' of a time series.}

\item{IDs_times}{optional argument used if 'grid' is NULL, vector of IDs associated with each start and end time.}
}
\value{
a data.table with values on a discrete time grids
}
\description{
Put Variable on time table
}
