% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PutOnTimeTable}
\alias{PutOnTimeTable}
\title{Put Variable on time table}
\usage{
PutOnTimeTable(
  values,
  IDs,
  datetimes = NULL,
  var_name = NULL,
  duration_hr = NA,
  sample_interval_minute = 30,
  aggregration_type = "latest",
  Template = NULL,
  start_times = NULL,
  end_times = NULL,
  IDs_times = NULL
)
}
\arguments{
\item{values}{a vector of values to be added to the time table.}

\item{IDs}{a vector of IDs associated with each value in values.}

\item{datetimes}{a vector of date-time objects associated with each value in values.}

\item{var_name}{a scalar string of the name of the variable to be added.}

\item{duration_hr}{a scalar integer denoting the number of hours for which a variable's value can be carried forward before it is considered missing data.}

\item{sample_interval_minute}{a scalar integer denoting the number of minutes between each time grid point in the time table.}

\item{aggregration_type}{how to handle multiple measurements for same time point; 'latest' or 'mean'}

\item{Template}{the time grid 'Template' precomputed from GetTimeTableTemplate()}

\item{start_times}{optional argument used if 'Template' is NULL vector of date times denoting the 'end' of a time series.}

\item{end_times}{optional argument used if 'Template' is NULL, vector of date times denoting the 'start' of a time series.}

\item{IDs_times}{optional argument used if 'Template' is NULL, vector of IDs associated with each start and end time.}
}
\value{
a data.table with values on a discrete time grids
}
\description{
Put Variable on time table
}
